// src/app/dashboard/create-escrow/interfaces.ts
export interface Escrow {
    escrow_id: string;
    initiator: string;
    transaction_title: string;
    role: string;
    currency: string;
    inspection_period: string;
    item_name: string;
    price: string;
    item_category: string;
    item_description: string;
    shipping_method: string;
    shipping_fee_paid_by: string;
    escrow_status: string;
    created_on: string;
}